+++ Godot/mods/Lucy.LucysTools/lucys_menu.gd
@@ -7,11 +7,16 @@ func setup(manager):
 	get_node("%lucy_bbcode").pressed = manager.allow_bbcode
 	get_node("%lucy_punchback").pressed = manager.do_punchback
 	get_node("%lucy_servername").text = manager.custom_server_name
+	get_node("%lucy_servername_preview").bbcode_text = manager.custom_server_name + "'s Lobby"
 	get_node("%lucy_servermsg").text = manager.server_join_message
+	get_node("%lucy_servermsg_preview").bbcode_text = manager.server_join_message
 	get_node("%lucy_fpackets").value = manager.frame_packets
 	get_node("%lucy_bpackets").value = manager.bulk_packets
 	get_node("%lucy_binterval").value = manager.bulk_interval
 	get_node("%lucy_finterval").value = manager.full_interval
+	
+	get_node("%lucy_chatcolor_bool").pressed = manager.custom_color_enabled
+	get_node("%lucy_chatcolor").color = Color(manager.custom_color)
 	
 	update()
 
@@ -22,10 +27,11 @@ func _ready():
 	print("[LUCY] Menu Ready")
 	
 	get_node("%lucy_bbcode").disabled = MANAGER.host_required and not Network.GAME_MASTER 
-	get_node("%lucy_raincloud").disabled = (MANAGER.host_required and not Network.GAME_MASTER) or not MANAGER.ingame
-	get_node("%lucy_meteor").disabled = (MANAGER.host_required and not Network.GAME_MASTER) or not MANAGER.ingame
-	get_node("%lucy_freezerain").disabled = (MANAGER.host_required and not Network.GAME_MASTER) or not MANAGER.ingame
-	get_node("%lucy_clearrain").disabled = (MANAGER.host_required and not Network.GAME_MASTER) or not MANAGER.ingame
+	get_node("%lucy_raincloud").disabled = not Network.GAME_MASTER or not MANAGER.ingame
+	get_node("%lucy_meteor").disabled = not Network.GAME_MASTER or not MANAGER.ingame
+	get_node("%lucy_freezerain").disabled = not Network.GAME_MASTER or not MANAGER.ingame
+	get_node("%lucy_clearrain").disabled = not Network.GAME_MASTER or not MANAGER.ingame
+	get_node("%lucy_clearmeteor").disabled = not Network.GAME_MASTER or not MANAGER.ingame
 
 func _input(event):
 	if event is InputEventKey and event.scancode == KEY_F5 && event.pressed:
@@ -45,8 +51,10 @@ func _on_lucy_bbcode_toggled(button_pressed):
 func _on_lucy_punchback_toggled(button_pressed):
 	MANAGER.do_punchback =  button_pressed
 func _on_lucy_servername_text_changed(new_text):
+	get_node("%lucy_servername_preview").bbcode_text = new_text + "'s Lobby"
 	MANAGER.custom_server_name = new_text
 func _on_lucy_servermsg_text_changed(new_text):
+	get_node("%lucy_servermsg_preview").bbcode_text = new_text
 	MANAGER.server_join_message = new_text
 func _on_lucy_fpackets_value_changed(value):
 	MANAGER.frame_packets = value
@@ -56,6 +64,10 @@ func _on_lucy_binterval_value_changed(value):
 	MANAGER.bulk_interval = value
 func _on_lucy_finterval_value_changed(value):
 	MANAGER.full_interval = value
+func _on_lucy_chatcolor_bool_toggled(button_pressed):
+	MANAGER.custom_color_enabled = button_pressed
+func _on_lucy_chatcolor_color_changed(color):
+	MANAGER.custom_color = color
 
 func _on_lucy_raincloud_pressed():
 	if not MANAGER.ingame: return
@@ -67,6 +79,7 @@ func _on_lucy_raincloud_pressed():
 
 func _on_lucy_meteor_pressed():
 	if not MANAGER.ingame: return
+	if get_tree().get_nodes_in_group("meteor").size() > 10: return
 	print("[LUCY] Spawning meteor")
 	var player_pos = MANAGER.get_player().global_transform.origin
 	var dist = INF
@@ -98,3 +111,10 @@ func _on_lucy_clearchat_pressed():
 	Network.GAMECHAT = ""
 	Network.LOCAL_GAMECHAT = ""
 	Network.emit_signal("_chat_update")
+
+func _on_lucy_clearmeteor_pressed():
+	if not MANAGER.ingame or not Network.GAME_MASTER: return
+	print("[LUCY] Clearing meteor")
+	for meteor in get_tree().get_nodes_in_group("meteor"):
+		meteor._deinstantiate(true)
+