+++ CHANGELOG.md
@@ -7,3 +7,13 @@ initial release
 - BBCode controlled by server host
 - Spawning requires host
 - Clear chat button
+
+0.2.1
+----
+- fix spawning oops
+
+0.3
+----
+- log settings
+- custom color
+- server name & join message previews